skip books that are already downloaded
Moved the `pdf_file_name` declaration to the top of the `try` block and put everything else inside `if not os.path.exists(pdf_file_name):`

+++ springer_book_scraper.py
@@ -11,13 +11,14 @@ BASE_DIR = '.'
 
 def download_pdfs_from_url(book_url, book_title, book_topic):
     try:
-        r = requests.get(book_url)
-        url_base = r.url
-        codigo_libro = url_base.split('/')[-1]
-        url_pdf = URL_PDF.format(codigo_libro = codigo_libro)
-        pdf_file = requests.get(url_pdf)
         pdf_file_name = '{base_dir}/{book_topic}/{book_title}.pdf'.format(book_topic=book_topic, book_title=book_title, base_dir=BASE_DIR)
-        open(pdf_file_name, 'wb').write(pdf_file.content)
+        if not os.path.exists(pdf_file_name):
+            r = requests.get(book_url)
+            url_base = r.url
+            codigo_libro = url_base.split('/')[-1]
+            url_pdf = URL_PDF.format(codigo_libro = codigo_libro)
+            pdf_file = requests.get(url_pdf)
+            open(pdf_file_name, 'wb').write(pdf_file.content)
     except:
         print("\n Error downloading: {book_topic} - {book_title}".format(book_topic=book_topic, book_title=book_title))
